+++ app/Http/Controllers/Admin/Sale/SalePOsController.php
@@ -33,6 +33,9 @@ class SalePOsController extends Controller
 
     public function index(Request $request)
     {
+      if (!auth()->user()->can('sale_pos.view')) {
+            abort(403, 'Unauthorized action.');
+        }
       if ($request->ajax()) {
             $q=Transaction::query();
             if (!empty(request()->input('sale_type'))) {
@@ -50,6 +53,10 @@ class SalePOsController extends Controller
             if (!empty(request()->input('created_by'))) {
                 $q=$q->where('created_by',request()->input('created_by'));
             }
+
+             if (!auth()->user()->hasRole('Super Admin')) {
+                $q=$q->where('hidden',false);
+            }
             $q =$q->where('transaction_type','Sale');
             $document =$q->get();
 
@@ -62,10 +69,18 @@ class SalePOsController extends Controller
                   return $model->client?$model->client->name:'';
                  })
                  ->editColumn('paid', function ($model) {
-                  return $model->payment()->where('type','Credit')->sum('amount');
+                    if (auth()->user()->can("view_sale.sale_paid")) {
+                    return $model->payment()->where('type','Credit')->sum('amount');
+                  }else{
+                    return 'N/A';
+                  }
                  })
                 ->editColumn('due', function ($model) {
-                  return $model->net_total-($model->payment()->where('type','Credit')->sum('amount'));
+                    if (auth()->user()->can("view_sale.sale_due")) {
+                    return $model->net_total-($model->payment()->where('type','Credit')->sum('amount'));
+                    }else{
+                        return 'N/A';
+                    }
                  })
                  ->editColumn('payment_status', function ($model) {
                    if ($model->payment_status=='paid') {
@@ -93,7 +108,10 @@ class SalePOsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function create()
-    {
+    { 
+        if (!auth()->user()->can('sale_pos.create')) {
+            abort(403, 'Unauthorized action.');
+        }
         $brands =Brand::select('id','name')->get();
         $categories=Category::all();
         return view('admin.salePos.create',compact('brands','categories'));
@@ -101,6 +119,9 @@ class SalePOsController extends Controller
 
     public function sale_add()
     {
+     if (!auth()->user()->can('sale_pos.create')) {
+            abort(403, 'Unauthorized action.');
+        }
        return view('admin.salePos.sale_add'); 
     }
 
@@ -112,6 +133,9 @@ class SalePOsController extends Controller
      */
     public function store(Request $request)
     {
+        if (!auth()->user()->can('sale_pos.create')) {
+            abort(403, 'Unauthorized action.');
+        }
         $input = $request->except('_token','variation');
 
          if (empty($request->input('date'))) {
@@ -195,6 +219,9 @@ class SalePOsController extends Controller
      */
     public function show($id)
     {
+     if (!auth()->user()->can('sale_pos.create')) {
+            abort(403, 'Unauthorized action.');
+        }
         $model =Transaction::findOrfail($id);
         return view('admin.salePos.show',compact('model'));
     }
@@ -230,6 +257,9 @@ class SalePOsController extends Controller
      */
     public function destroy($id)
     {
+    if (!auth()->user()->can('sale_pos.delete')) {
+            abort(403, 'Unauthorized action.');
+        }
        $transaction =Transaction::find($id);
        //Sale Related
        $transaction->sell_lines()->delete();
@@ -287,6 +317,9 @@ class SalePOsController extends Controller
                     $query->orWhere('products.sub_category_id', $category_id);
                 });
           }
+          if (!auth()->user()->hasRole('Super Admin')) {
+                $products->where('variations.hidden',false);
+            }
            // $products->where(function ($query) use ($brand_id) {
            //          $query->where('VBD.brand_id', $brand_id);
                     
@@ -382,6 +415,9 @@ class SalePOsController extends Controller
 
     public function view($id)
     {
+     if (!auth()->user()->can('sale_pos.view')) {
+            abort(403, 'Unauthorized action.');
+        }
         $model =Transaction::find($id);
         return view('admin.salePos.partials.view',compact('model'));
     }
